pyproject.toml
[project]
name = "terminai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "genson>=1.3.0",
    "huggingface-hub>=1.0.1",
    "llama-cpp-python>=0.3.16",
]
